rename types.ts to parser.ts to not have a generic file
