--- packages/next-swr-endpoints/src/types.ts
@@ -1,3 +1,0 @@
-export type Parser<Into> =
-  | { parse(arg: unknown): Into }
-  | { parseAsync(arg: unknown): Promise<Into> };
